+++ src/Physics/HadronTensors/LinkDef.h
@@ -1,0 +1,15 @@
+#ifdef __CINT__
+
+#pragma link off all globals;
+#pragma link off all classes;
+#pragma link off all functions;
+
+#pragma link C++ namespace genie;
+
+#pragma link C++ class genie::HadronTensorI;
+#pragma link C++ class genie::ValenciaHadronTensorI;
+#pragma link C++ class genie::TabulatedValenciaHadronTensor;
+
+#pragma link C++ class genie::HadronTensorPool;
+
+#endif
